kjun.ai.kr/ai.kjun.ai.kr/mlservice/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn  # datasets 포함 (from sklearn import datasets)
lightgbm
requests  # HTTP 요청용 (카카오 맵 API)

#엑셀파일읽기기
openpyxl  # Excel 파일 읽기용 (.xlsx)
xlrd  # Excel 파일 읽기용 (.xls)

#디버깅 도구
icecream

#환경변수관리
python-dotenv  # .env 파일 로드용
folium>=0.12.0  # 지도 시각화용

#시각화
matplotlib  # 히트맵 생성용
seaborn  # 히트맵 생성용

# NLP 관련 패키지
nltk  # 자연어 처리
wordcloud  # 워드클라우드 생성
sentencepiece  # 토크나이저
konlpy  # 한국어 NLP (Java 필요)
opencv-python  # 이미지 처리
JPype1  # Java 연동 (KoNLPy용)
Pillow  # 이미지 처리